pyproject.toml
[project]
name = "ai-improve-issue"
version = "0.2.0"
description = "A feature that automatically generates specific text following an appropriate template and presents it as a comment, after the user writes 1-2 lines describing their objective in a simple blank Issue."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10.0",
    "pydantic-ai>=0.0.14",
    "pydantic-graph>=0.4.1",
    "voyageai>=0.2.3",
    "qdrant-client==1.16.*",
    "langchain-text-splitters>=0.3.0",
    "pyyaml>=6.0",
]

[project.scripts]
ai-improve-issue = "ai_improve_issue.main:main"

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"
